+++ firmware/src/modules/rgb.rs
@@ -34,11 +34,16 @@ impl RgbMod {
     ) -> Self {
         count_down.start(FRAME_TIME.millis());
 
+        let mut default_rgb_profile = RgbProfile::default_device_profile();
+        DEFAULT_RGB_PROFILE.with_lock(|p| {
+            default_rgb_profile = p.1.clone();
+        });
+
         RgbMod {
             ws: ws,
             buffer: [(0, 0, 0).into(); RGB_LEN],
             timer: count_down,
-            rgb_mode: DEFAULT_RGB_PROFILE,
+            rgb_mode: default_rgb_profile,
         }
     }